[Modify] 스타일 작업을 위한 정돈
